dblens-mvp-starter/services/ingester/requirements.txt
psycopg[binary]==3.2.1
pandas==2.2.2
pyarrow==16.1.0
requests==2.32.3
python-magic==0.4.27